实现简历pdf下载一致
