+++ src/views/CreateResume.vue
@@ -119,10 +119,8 @@ import metadataInstance from '@/models/metadata_model.js';
 import ChatgptModel from '@/models/chatgpt_model.js';
 import { waveform } from 'ldrs';
 import { resumeModel } from '@/models/resume_model.js';
-import apiClient from '@/api/axios';
 import { useToast } from 'vue-toastification';
-import JSZip from 'jszip';
-import { saveAs } from 'file-saver';
+import { downloadResume } from '@/utils/pdfdownload';
 
 waveform.register();
 const chatgptInstance = ChatgptModel.getInstance();
@@ -281,69 +279,21 @@ export default {
     /**
      * 下载截图
      */
-    handleCaptureAndSaveScreenshot() {
+    async handleCaptureAndSaveScreenshot() {
       if (this.isDownloading) return
       this.isDownloading = true
-
-      // 提示下载开始
-      this.toast.success('简历下载中')
-
-      // 获取当前简历ID
-      const resumeId = this.$route.params.resumeId
-
-      // 使用新的截图接口获取实时截图
-      apiClient
-        .post('/pic/scf-screenshot', {
-          resumeId,
-          templateType: this.templateType,
-          color: this.color,
-        })
-        .then(async (response) => {
-          const urls = response.data.data.screenshotUrls
-          if (response.data.code === 20009 && Array.isArray(urls) && urls.length) {
-            if (urls.length === 1) {
-              const link = document.createElement('a')
-              link.href = urls[0]
-              link.download = `${response.data.data.resumeId}.png`
-              document.body.appendChild(link)
-              link.click()
-              document.body.removeChild(link)
-            } else {
-              await this.downloadZip(urls, response.data.data.resumeId)
-            }
-          } else if (response.data.data.screenshotUrl) {
-            // 向后兼容旧接口
-            const link = document.createElement('a')
-            link.href = response.data.data.screenshotUrl
-            link.download = `${response.data.data.resumeId}.png`
-            document.body.appendChild(link)
-            link.click()
-            document.body.removeChild(link)
-          } else {
-            this.toast.error('下载失败，请重试')
-            console.error('获取简历截图失败:', response.data.message)
-          }
-        })
-        .catch((error) => {
+      
+      await downloadResume({
+        resumeId: this.$route.params.resumeId,
+        templateType: this.templateType,
+        color: this.color,
+      }).catch((error) => {
           console.error('下载简历截图时出错:', error)
           this.toast.error('下载失败，请重试')
         })
         .finally(() => {
           this.isDownloading = false
         })
-    },
-
-    async downloadZip(urls, baseName) {
-      const zip = new JSZip()
-      await Promise.all(
-        urls.map((url, index) =>
-          apiClient.get(url, { responseType: 'blob' }).then((res) => {
-            zip.file(`${baseName}_${index + 1}.png`, res.data)
-          })
-        )
-      )
-      const blob = await zip.generateAsync({ type: 'blob' })
-      saveAs(blob, `${baseName}.zip`)
     },
     /**
      * 编辑某个标题
