+++ src/views/HomeLogged.vue
@@ -259,9 +259,7 @@ import apiClient from '@/api/axios'
 import { waveform } from 'ldrs'
 import { useToast } from 'vue-toastification'
 import { Solar } from 'lunar-javascript'
-import JSZip from 'jszip'
-import { saveAs } from 'file-saver'
-import { PDFDocument } from 'pdf-lib'
+import { downloadResume as downloadResumeUtil } from '@/utils/pdfdownload'
 
 // ====== 新增：导入简历弹窗组件 ======
 import ImportResumeModal from '@/components/ImportResumeModal.vue'
@@ -613,111 +611,17 @@ export default {
         resume.showDropdown = !resume.showDropdown
       }
     },
-    async downloadPngUrlAspdf(pngUrl,fileName="简历") {
-      // 把图片取回来
-      const pngBytes = await fetch(pngUrl).then(res => res.arrayBuffer())
-      // 创建 PDF 文档
-      const pdfDoc = await PDFDocument.create()
-      const pngImage = await pdfDoc.embedPng(pngBytes)
-      const { width, height } = pngImage.scale(0.6)
-      const page = pdfDoc.addPage([width, height])
-      page.drawImage(pngImage, { x: 0, y: 0, width, height })
-      const pdfBytes = await pdfDoc.save()
-      // 下载 PDF
-      const blob = new Blob([pdfBytes], { type: 'application/pdf' })
-      const url = URL.createObjectURL(blob)
-      const a = document.createElement('a')
-      a.href = url
-      a.download = `${fileName}.pdf`
-      document.body.appendChild(a)
-      a.click()
-      document.body.removeChild(a)
-      URL.revokeObjectURL(url)
-    },
     async downloadResume(resume) {
       if (resume.isDownloading) return
       resume.isDownloading = true
       try {
-        this.toast.success('简历下载中')
-        const response = await apiClient.post('/pic/scf-screenshot', {
-          resumeId: resume.resumeId,
-          templateType: resume.templateType,
-          color: resume.color,
-        })
-
-        const urls = response.data.data.screenshotUrls
-        if (response.data.code === 20009 && Array.isArray(urls) && urls.length) {
-          if (urls.length === 1) {
-            await this.downloadPngUrlAspdf(urls[0])
-            /*const a = document.createElement('a')
-            a.href = urls[0]
-            a.download = `${resume.name || '简历'}.png`
-            document.body.appendChild(a)
-            a.click()
-            document.body.removeChild(a)*/
-          } else {
-            await this.downloadImagesAsPdf(urls, resume.name || '简历')
-          }
-        } else if (response.data.data.screenshotUrl) {
-          // 向后兼容旧接口
-          await this.downloadPngUrlAspdf(response.data.data.screenshotUrl)
-        } else {
-          this.toast.error('下载失败，请重试')
-        }
+        await downloadResumeUtil(resume)
       } catch (error) {
         console.error('下载失败:', error)
         this.toast.error('下载失败，请重试')
       } finally {
         resume.isDownloading = false
       }
-    },
-    async downloadImagesAsPdf(urls, fileName = "简历") {
-      // 创建一个 PDF 文档
-      const pdfDoc = await PDFDocument.create()
-      
-      for (let url of urls) {
-        // 下载图片数据
-        const imgBytes = await fetch(url).then(res => res.arrayBuffer())
-        
-        // 判断图片类型（这里只写 PNG / JPG）
-        let img
-        if (url.toLowerCase().endsWith('.png')) {
-          img = await pdfDoc.embedPng(imgBytes)
-        } else {
-          img = await pdfDoc.embedJpg(imgBytes)
-        }
-        
-        const { width, height } = img.scale(1)
-        // 每张图一页
-        const page = pdfDoc.addPage([width, height])
-        page.drawImage(img, { x: 0, y: 0, width, height })
-      }
-
-      // 导出 PDF
-      const pdfBytes = await pdfDoc.save()
-      const blob = new Blob([pdfBytes], { type: 'application/pdf' })
-      const pdfUrl = URL.createObjectURL(blob)
-
-      // 触发下载
-      const a = document.createElement('a')
-      a.href = pdfUrl
-      a.download = `${fileName}.pdf`
-      document.body.appendChild(a)
-      a.click()
-      document.body.removeChild(a)
-      URL.revokeObjectURL(pdfUrl)
-    },
-    async downloadZip(urls, baseName) {
-      const zip = new JSZip()
-      await Promise.all(
-        urls.map((url, index) =>
-          apiClient.get(url, { responseType: 'blob' }).then((res) => {
-            zip.file(`${baseName}_${index + 1}.png`, res.data)
-          })
-        )
-      )
-      const blob = await zip.generateAsync({ type: 'blob' })
-      saveAs(blob, `${baseName}.zip`)
     },
     renameResume(resume) {
       const newName = prompt('请输入新的简历名称', resume.name)
